# SimpleCardGame/WebApp/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "Simple Gates Game API",
        "description": "A simple representation of Simple Gates Game."
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/game": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Retrieves the current state of the memory game board.",
                "operationId": "game_get",
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "players": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Player"
                                    }
                                },
                                "deck": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Card"
                                    }
                                },
                                "player_turn": {
                                    "type": "string"
                                },
                                "selected_cards": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Card"
                                    }
                                },
                                "played_cards": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Card"
                                    }
                                },
                                "cards": {
                                    "type": "object",
                                    "properties": {
                                        "player_id": {
                                            "type": "string"
                                        },
                                        "cards": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/Card"
                                            }
                                        }
                                    }
                                },
                                "qubits": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Qubit"
                                    }
                                }
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/play": {
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Specifies a card to reveal and checks for a match to the previous.",
                "operationId": "game_play",
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "player_id",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "card_index",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "qubit_index",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "player_id": {
                                    "type": "string"
                                },
                                "qubits": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Qubit"
                                    }
                                }
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/new": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Initializes a new game board",
                "operationId": "game_new",
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "players",
                        "in": "query",
                        "required": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/NewPlayer"
                        }
                    },
                    {
                        "name": "card_types",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "deprecated": false
            }
        }
    },
    "definitions": {
        "Card": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "player_id": {
                    "type": "string"
                }
            }
        },
        "CardType": {
            "type": "string"
        },
        "CardPlay": {
            "type": "object",
            "properties": {
                "card_index": {
                    "format": "int32",
                    "type": "integer"
                },
                "qubit_index": {
                    "format": "int32",
                    "type": "integer"
                }
            }
        },
        "NewPlayer": {
            "type": "string"
        },
        "Player": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "score": {
                    "format": "int32",
                    "type": "integer"
                },
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Card"
                    }
                }
            }
        },
        "Qubit": {
            "type": "object",
            "properties": {
                "value1": {
                    "type": "integer",
                    "format": "int32"
                },
                "value2": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        }
    }
}